回答機能実装

+++ app/Http/Controllers/QuestionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Http\Requests\QuestionRequest;
+use App\Http\Requests\AnswerRequest;
 use App\Question;
 use App\Review;
 use App\Answer;
@@ -15,7 +16,7 @@ class QuestionController extends Controller
         return view('questions/create')->with(['review' => $review]);
     }
     
-    public function store(Review $review, Question $question, Request $request)
+    public function store(Review $review, Question $question, QuestionRequest $request)
     {
         $input = $request['question'];
         $input += ['user_id' => $request->user()->id];
@@ -27,17 +28,39 @@ class QuestionController extends Controller
     
     public function delete(Review $review, Question $question)
     {
-        $question->delete();
-        return redirect('/reviews/' . $review->id);
+        if(\Auth::user()->id == $question->user_id){
+            $question->delete();
+            return redirect('/reviews/' . $review->id);
+        }else{
+            abort(403, '権限がありません');
+        }
     }
     
-    public function answercreate(Question $question)
+    public function answercreate(Review $review, Question $question)
     {
-        return view('questions/answercreate')->with(['question' => $question]);    
+        return view('questions/answercreate')->with(['question' => $question])->with(['review' => $review]);    
     }
     
-    public function answerstore(Question $question, Answer $answer, QuestionRequest $request)
+    public function answerstore(Review $review, Question $question, Answer $answer, AnswerRequest $request)
     {
-        
+        if(\Auth::user()->id == $review->user_id){
+            $input = $request['answer'];
+            $input += ['question_id' => $question->id];
+            
+            $answer->fill($input)->save();
+            return redirect('/reviews/' . $review->id);
+        }else{
+            abort(403, '権限がありません');
+        }
+    }
+    
+    public function answerdelete(Review $review, Question $question, Answer $answer)
+    {
+        if(\Auth::user()->id == $review->user_id){
+            $answer->delete();
+            return redirect('/reviews/' . $review->id);
+        }else{
+            abort(403, '権限がありません');
+        }
     }
 }
